rework frontend for maintenances (factorize)
